requirements.txt
numpy
Pillow
ttkbootstrap
requests
